refactor: Improve deposit liquidity functionality and UI
This commit includes several improvements to the deposit liquidity functionality:

- Refactor deposit-liquidity.ts to use createRawTx with proper fee handling
- Add electrumFetchNonGlittrUtxos and addFeeToTx for consistent transaction building
- Create specialized LiquidityDepositModal component with confirmation state
- Update both swap and create-pool pages to use the new modal component
- Modernize the original pool creation modal with consistent styling
- Fix typo in "Deposit" throughout the codebase
- Improve transaction feedback and explorer links in all modals
- Add proper transaction confirmation state indicators

These changes make the deposit liquidity flow more consistent across the application
and provide better user feedback during transaction lifecycle.
